move login and register controllers to a separate file authController.ts
